Move color types to subpackage. #222
