+++ eventdiscovery-sdk/src/main/java/com/schedjoules/eventdiscovery/framework/microfragments/eventdetails/fragments/views/EventHeaderView.java
@@ -27,9 +27,9 @@ import com.schedjoules.client.eventsdiscovery.Event;
 import com.schedjoules.eventdiscovery.R;
 import com.schedjoules.eventdiscovery.databinding.SchedjoulesDetailsHeaderBinding;
 import com.schedjoules.eventdiscovery.framework.common.StatusBar;
-import com.schedjoules.eventdiscovery.framework.utils.AttributeColor;
 import com.schedjoules.eventdiscovery.framework.utils.SchedJoulesLinks;
-import com.schedjoules.eventdiscovery.framework.utils.Transparent;
+import com.schedjoules.eventdiscovery.framework.utils.colors.AttributeColor;
+import com.schedjoules.eventdiscovery.framework.utils.colors.Transparent;
 import com.schedjoules.eventdiscovery.framework.utils.smartview.SmartView;
