+++ eventdiscovery-sdk/src/main/java/com/schedjoules/eventdiscovery/framework/utils/colors/Color.java
@@ -15,7 +15,7 @@
  * limitations under the License.
  */
 
-package com.schedjoules.eventdiscovery.framework.utils;
+package com.schedjoules.eventdiscovery.framework.utils.colors;
 
 import android.support.annotation.ColorInt;
